+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 2.3.0
 
+### Breaking changes
+
+* `RealmResults.distinct()` returns a new `RealmResults` object instead of filtering on the original object.
+
 ### Object Server API Changes (In Beta)
 
 * Add a default `UserStore` based on the Realm Object Store (`ObjectStoreUserStore`).